requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.1
opencv-python==4.6.0.66
Pillow==9.3.0
pyzbar==0.1.9
Werkzeug==2.2.2
waitress==2.1.2